[GH-4] Move teacher workload endpoint to Workload controller
